made header server component
